+++ CorsixTH/Src/th_lua_anims.cpp
@@ -112,7 +112,7 @@ static int l_anims_loadcustom(lua_State *L)
 static int l_anims_getanims(lua_State *L)
 {
     THAnimationManager* pAnims = luaT_testuserdata<THAnimationManager>(L);
-    int iTileSize = luaL_checkinteger(L, 2);
+    int iTileSize = static_cast<int>(luaL_checkinteger(L, 2));
     const char *pName = luaL_checkstring(L, 3);
 
     const AnimationStartFrames &oFrames = pAnims->getNamedAnimations(pName, iTileSize);
@@ -127,7 +127,7 @@ static int l_anims_getanims(lua_State *L)
 static int l_anims_getfirst(lua_State *L)
 {
     THAnimationManager* pAnims = luaT_testuserdata<THAnimationManager>(L);
-    int iAnim = luaL_checkinteger(L, 2);
+    int iAnim = static_cast<int>(luaL_checkinteger(L, 2));
 
     lua_pushinteger(L, pAnims->getFirstFrame(iAnim));
     return 1;
@@ -136,7 +136,7 @@ static int l_anims_getfirst(lua_State *L)
 static int l_anims_getnext(lua_State *L)
 {
     THAnimationManager* pAnims = luaT_testuserdata<THAnimationManager>(L);
-    int iFrame = luaL_checkinteger(L, 2);
+    int iFrame = static_cast<int>(luaL_checkinteger(L, 2));
 
     lua_pushinteger(L, pAnims->getNextFrame(iFrame));
     return 1;
@@ -168,7 +168,7 @@ static int l_anims_set_marker(lua_State *L)
 {
     THAnimationManager* pAnims = luaT_testuserdata<THAnimationManager>(L);
     lua_pushboolean(L, pAnims->setFrameMarker(luaL_checkinteger(L, 2),
-        luaL_checkinteger(L, 3), luaL_checkinteger(L, 4)) ? 1 : 0);
+        static_cast<int>(luaL_checkinteger(L, 3)), static_cast<int>(luaL_checkinteger(L, 4))) ? 1 : 0);
     return 1;
 }
 
@@ -176,7 +176,7 @@ static int l_anims_set_secondary_marker(lua_State *L)
 {
     THAnimationManager* pAnims = luaT_testuserdata<THAnimationManager>(L);
     lua_pushboolean(L, pAnims->setFrameSecondaryMarker(luaL_checkinteger(L, 2),
-        luaL_checkinteger(L, 3), luaL_checkinteger(L, 4)) ? 1 : 0);
+        static_cast<int>(luaL_checkinteger(L, 3)), static_cast<int>(luaL_checkinteger(L, 4))) ? 1 : 0);
     return 1;
 }
 
@@ -186,9 +186,9 @@ static int l_anims_draw(lua_State *L)
     THRenderTarget* pCanvas = luaT_testuserdata<THRenderTarget>(L, 2);
     size_t iFrame = luaL_checkinteger(L, 3);
     THLayers_t* pLayers = luaT_testuserdata<THLayers_t>(L, 4, luaT_upvalueindex(2));
-    int iX = luaL_checkinteger(L, 5);
-    int iY = luaL_checkinteger(L, 6);
-    int iFlags = luaL_optinteger(L, 7, 0);
+    int iX = static_cast<int>(luaL_checkinteger(L, 5));
+    int iY = static_cast<int>(luaL_checkinteger(L, 6));
+    int iFlags = static_cast<int>(luaL_optinteger(L, 7, 0));
 
     pAnims->drawFrame(pCanvas, iFrame, *pLayers, iX, iY, iFlags);
 
@@ -302,7 +302,7 @@ static int l_anim_get_frame(lua_State *L)
 static int l_anim_set_crop(lua_State *L)
 {
     THAnimation* pAnimation = luaT_testuserdata<THAnimation>(L);
-    pAnimation->setCropColumn(luaL_checkinteger(L, 2));
+    pAnimation->setCropColumn(static_cast<int>(luaL_checkinteger(L, 2)));
     lua_settop(L, 1);
     return 1;
 }
@@ -325,7 +325,7 @@ static int l_anim_set_anim(lua_State *L)
     if(lua_isnoneornil(L, 4))
         pAnimation->setFlags(0);
     else
-        pAnimation->setFlags(luaL_checkinteger(L, 4));
+        pAnimation->setFlags(static_cast<uint32_t>(luaL_checkinteger(L, 4)));
 
     pAnimation->setAnimation(pManager, iAnim);
     lua_settop(L, 2);
@@ -343,7 +343,7 @@ static int l_anim_set_morph(lua_State *L)
 
     unsigned int iDurationFactor = 1;
     if(!lua_isnoneornil(L, 3) && luaL_checkinteger(L, 3) > 0)
-        iDurationFactor = luaL_checkinteger(L, 3);
+        iDurationFactor = static_cast<unsigned int>(luaL_checkinteger(L, 3));
 
     pAnimation->setMorphTarget(pMorphTarget, iDurationFactor);
     lua_settop(L, 2);
@@ -354,7 +354,7 @@ static int l_anim_set_morph(lua_State *L)
 
 static int l_anim_set_drawable_layer(lua_State *L)
 {
-    lastLayer = luaL_checkinteger(L, 2);
+    lastLayer = static_cast<int>(luaL_checkinteger(L, 2));
     return 1;
 }
 
@@ -381,7 +381,7 @@ static int l_anim_set_tile(lua_State *L)
     else
     {
         THMap* pMap = luaT_testuserdata<THMap>(L, 2);
-        THMapNode* pNode = pMap->getNode(luaL_checkinteger(L, 3) - 1, luaL_checkinteger(L, 4) - 1);
+        THMapNode* pNode = pMap->getNode(static_cast<int>(luaL_checkinteger(L, 3) - 1), static_cast<int>(luaL_checkinteger(L, 4) - 1));
         if(pNode)
             pAnimation->attachToTile(pNode, lastLayer);
 
@@ -446,7 +446,7 @@ template <typename T>
 static int l_anim_set_flag(lua_State *L)
 {
     T* pAnimation = luaT_testuserdata<T>(L);
-    pAnimation->setFlags(luaL_checkinteger(L, 2));
+    pAnimation->setFlags(static_cast<uint32_t>(luaL_checkinteger(L, 2)));
 
     lua_settop(L, 1);
     return 1;
@@ -456,7 +456,7 @@ template <typename T>
 static int l_anim_set_flag_partial(lua_State *L)
 {
     T* pAnimation = luaT_testuserdata<T>(L);
-    uint32_t iFlags = luaL_checkinteger(L, 2);
+    uint32_t iFlags = static_cast<uint32_t>(luaL_checkinteger(L, 2));
     if(lua_isnone(L, 3) || lua_toboolean(L, 3))
     {
         pAnimation->setFlags(pAnimation->getFlags() | iFlags);
@@ -503,7 +503,7 @@ static int l_anim_set_position(lua_State *L)
 {
     T* pAnimation = luaT_testuserdata<T>(L);
 
-    pAnimation->setPosition(luaL_checkinteger(L, 2), luaL_checkinteger(L, 3));
+    pAnimation->setPosition(static_cast<int>(luaL_checkinteger(L, 2)), static_cast<int>(luaL_checkinteger(L, 3)));
 
     lua_settop(L, 1);
     return 1;
@@ -524,7 +524,7 @@ static int l_anim_set_speed(lua_State *L)
 {
     T* pAnimation = luaT_testuserdata<T>(L);
 
-    pAnimation->setSpeed(luaL_optinteger(L, 2, 0), luaL_optinteger(L, 3, 0));
+    pAnimation->setSpeed(static_cast<int>(luaL_optinteger(L, 2, 0)), static_cast<int>(luaL_optinteger(L, 3, 0)));
 
     lua_settop(L, 1);
     return 1;
@@ -535,7 +535,7 @@ static int l_anim_set_layer(lua_State *L)
 {
     T* pAnimation = luaT_testuserdata<T>(L);
 
-    pAnimation->setLayer(luaL_checkinteger(L, 2), luaL_optinteger(L, 3, 0));
+    pAnimation->setLayer(static_cast<int>(luaL_checkinteger(L, 2)), static_cast<int>(luaL_optinteger(L, 3, 0)));
 
     lua_settop(L, 1);
     return 1;
@@ -605,7 +605,7 @@ static int l_anim_draw(lua_State *L)
 {
     T* pAnimation = luaT_testuserdata<T>(L);
     THRenderTarget* pCanvas = luaT_testuserdata<THRenderTarget>(L, 2);
-    pAnimation->draw(pCanvas, luaL_checkinteger(L, 3), luaL_checkinteger(L, 4));
+    pAnimation->draw(pCanvas, static_cast<int>(luaL_checkinteger(L, 3)), static_cast<int>(luaL_checkinteger(L, 4)));
     lua_settop(L, 1);
     return 1;
 }
@@ -625,7 +625,7 @@ static int l_srl_append(lua_State *L)
 {
     THSpriteRenderList *pSrl = luaT_testuserdata<THSpriteRenderList>(L);
     pSrl->appendSprite(luaL_checkinteger(L, 2),
-                       luaL_checkinteger(L, 3), luaL_checkinteger(L, 4));
+                       static_cast<int>(luaL_checkinteger(L, 3)), static_cast<int>(luaL_checkinteger(L, 4)));
     lua_settop(L, 1);
     return 1;
 }
@@ -633,7 +633,7 @@ static int l_srl_append(lua_State *L)
 static int l_srl_set_lifetime(lua_State *L)
 {
     THSpriteRenderList *pSrl = luaT_testuserdata<THSpriteRenderList>(L);
-    pSrl->setLifetime(luaL_checkinteger(L, 2));
+    pSrl->setLifetime(static_cast<int>(luaL_checkinteger(L, 2)));
     lua_settop(L, 1);
     return 1;
 }
